Rework cover image drawinging
Adds extra controls for adjusting the cover image (scale and position).
This is implemented with both on-screen controls and mouse
input events. (Click and drag to reposition, use wheel to scale/zoom).

This change fixes image stretching to fit the canvas, and also
adds an auto-fit feature that scales and repositions the image
on load (or with a on-screen button press).

Several options have been changed to update the preview
instantly. To accommodate this rendering is now done on animation
frame.

Also includes configurable clear color, which is set to black by default.